refactor: remove v2 folders in operate components and utils
