+++ operate/client/src/App/Processes/MigrationView/TopPanel/Diagrams/SourceDiagram.test.tsx
